+++ tests/Commands/__snapshots__/ResourceMakeCommandTest__it_generated_correct_file_with_content__1.php
@@ -1,0 +1,20 @@
+<?php return '<?php
+
+namespace Modules\\Blog\\Transformers;
+
+use Illuminate\\Http\\Resources\\Json\\Resource;
+
+class PostsTransformer extends Resource
+{
+    /**
+     * Transform the resource into an array.
+     *
+     * @param  \\Illuminate\\Http\\Request
+     * @return array
+     */
+    public function toArray($request)
+    {
+        return parent::toArray($request);
+    }
+}
+';